# backend/docs/swagger.yaml
definitions:
...
  github_com_besart951_go_infra_link_backend_internal_handler_dto_auth.AuthResponse:
    properties:
      access_token_expires_at:
        type: string
      csrf_token:
        type: string
      refresh_token_expires_at:
        type: string
      user:
        $ref: '#/definitions/github_com_besart951_go_infra_link_backend_internal_handler_dto_auth.AuthUserResponse'
    type: object
  github_com_besart951_go_infra_link_backend_internal_handler_dto_auth.AuthUserResponse:
    properties:
      created_at:
        type: string
      disabled_at:
        type: string
      email:
        type: string
      failed_login_attempts:
        type: integer
      first_name:
        type: string
      id:
        type: string
      is_active:
        type: boolean
      last_login_at:
        type: string
      last_name:
        type: string
      locked_until:
        type: string
      permissions:
        items:
          type: string
        type: array
      role:
        type: string
      updated_at:
        type: string
    type: object
  github_com_besart951_go_infra_link_backend_internal_handler_dto_auth.ErrorResponse:
    properties:
      error:
        type: string
      fields:
        additionalProperties:
          type: string
        type: object
      message:
        type: string
    type: object
  github_com_besart951_go_infra_link_backend_internal_handler_dto_auth.LoginRequest:
    properties:
      email:
        type: string
      password:
        minLength: 8
        type: string
    required:
    - email
    - password
    type: object
...
  github_com_besart951_go_infra_link_backend_internal_handler_dto_facility.BacnetObjectResponse:
    properties:
      alarm_definition_id:
        type: string
      alarm_type_id:
        type: string
      created_at:
        type: string
      description:
        type: string
      field_device_id:
        type: string
      gms_visible:
        type: boolean
      hardware_quantity:
        type: integer
      hardware_type:
        type: string
      id:
        type: string
      notification_class_id:
        type: string
      optional:
        type: boolean
      software_number:
        type: integer
      software_reference_id:
        type: string
      software_type:
        type: string
      state_text_id:
        type: string
      text_fix:
        type: string
      text_individual:
        type: string
      updated_at:
        type: string
    type: object
...
  github_com_besart951_go_infra_link_backend_internal_handler_dto_facility.CreateBacnetObjectRequest:
    properties:
      alarm_definition_id:
        type: string
      alarm_type_id:
        type: string
      description:
        type: string
      field_device_id:
        type: string
      gms_visible:
        type: boolean
      hardware_quantity:
        maximum: 255
        minimum: 0
        type: integer
      hardware_type:
        type: string
      notification_class_id:
        type: string
      object_data_id:
        type: string
      optional:
        type: boolean
      software_number:
        maximum: 65535
        minimum: 0
        type: integer
      software_reference_id:
        type: string
      software_type:
        type: string
      state_text_id:
        type: string
      text_fix:
        maxLength: 250
        type: string
      text_individual:
        maxLength: 250
        type: string
    required:
    - software_number
    - software_type
    - text_fix
    type: object
...
  github_com_besart951_go_infra_link_backend_internal_handler_dto_facility.ErrorResponse:
    properties:
      error:
        type: string
      fields:
        additionalProperties:
          type: string
        type: object
      message:
        type: string
    type: object
...
  github_com_besart951_go_infra_link_backend_internal_handler_dto_facility.UpdateBacnetObjectRequest:
    properties:
      alarm_definition_id:
        type: string
      alarm_type_id:
        type: string
      description:
        type: string
      field_device_id:
        type: string
      gms_visible:
        type: boolean
      hardware_quantity:
        maximum: 255
        minimum: 1
        type: integer
      hardware_type:
        type: string
      notification_class_id:
        type: string
      object_data_id:
        type: string
      optional:
        type: boolean
      software_number:
        maximum: 65535
        minimum: 0
        type: integer
      software_reference_id:
        type: string
      software_type:
        type: string
      state_text_id:
        type: string
      text_fix:
        maxLength: 250
        type: string
      text_individual:
        maxLength: 250
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /api/v1/auth/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: Login data
        in: body
        name: login
        required: true
        schema:
          $ref: '#/definitions/github_com_besart951_go_infra_link_backend_internal_handler_dto_auth.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_besart951_go_infra_link_backend_internal_handler_dto_auth.AuthResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_besart951_go_infra_link_backend_internal_handler_dto_auth.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_besart951_go_infra_link_backend_internal_handler_dto_auth.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_besart951_go_infra_link_backend_internal_handler_dto_auth.ErrorResponse'
      summary: Login
      tags:
      - auth
...
  /api/v1/auth/me:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_besart951_go_infra_link_backend_internal_handler_dto_auth.AuthUserResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_besart951_go_infra_link_backend_internal_handler_dto_auth.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_besart951_go_infra_link_backend_internal_handler_dto_auth.ErrorResponse'
      summary: Get current user
      tags:
      - auth
  /api/v1/auth/refresh:
    post:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_besart951_go_infra_link_backend_internal_handler_dto_auth.AuthResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_besart951_go_infra_link_backend_internal_handler_dto_auth.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_besart951_go_infra_link_backend_internal_handler_dto_auth.ErrorResponse'
      summary: Refresh access token
      tags:
      - auth
...
  /api/v1/facility/bacnet-objects:
    post:
      consumes:
      - application/json
      parameters:
      - description: Bacnet Object data
        in: body
        name: bacnet_object
        required: true
        schema:
          $ref: '#/definitions/github_com_besart951_go_infra_link_backend_internal_handler_dto_facility.CreateBacnetObjectRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_besart951_go_infra_link_backend_internal_handler_dto_facility.BacnetObjectResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_besart951_go_infra_link_backend_internal_handler_dto_facility.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/github_com_besart951_go_infra_link_backend_internal_handler_dto_facility.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_besart951_go_infra_link_backend_internal_handler_dto_facility.ErrorResponse'
      summary: Create a bacnet object (for field device or object data)
      tags:
      - facility-bacnet-objects
  /api/v1/facility/bacnet-objects/{id}:
    put:
      consumes:
      - application/json
      parameters:
      - description: Bacnet Object ID
        in: path
        name: id
        required: true
        type: string
      - description: Bacnet Object data
        in: body
        name: bacnet_object
        required: true
        schema:
          $ref: '#/definitions/github_com_besart951_go_infra_link_backend_internal_handler_dto_facility.UpdateBacnetObjectRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_besart951_go_infra_link_backend_internal_handler_dto_facility.BacnetObjectResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_besart951_go_infra_link_backend_internal_handler_dto_facility.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_besart951_go_infra_link_backend_internal_handler_dto_facility.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/github_com_besart951_go_infra_link_backend_internal_handler_dto_facility.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_besart951_go_infra_link_backend_internal_handler_dto_facility.ErrorResponse'
      summary: Update a bacnet object
      tags:
      - facility-bacnet-objects
...
swagger: "2.0"